Initial update of grammar
I've updated this grammar based on the clojure and commonlisp
libraries. Currently it can parse the entire emacs library with no
errors (aside from some obsolete and ethiopian files that include
non-assigned unicode categories). The next step will be refining the
highlighting to match as much as the standard emacs syntax. Also
need to identify efficient way to highlight built-in macros.